Extract Class Refactoring
