+++ src/main/java/com/bric/math/Equations.java
@@ -70,46 +70,6 @@ public class Equations {
         return v1 - v2;
     }
 
-    public static String toString(double[][] d) {
-        String s = "";
-        for (double[] doubles : d) {
-            s = s + toString(doubles) + "\n";
-        }
-        return s.trim();
-    }
-
-    public static String toString(double[] d) {
-        String s = "[";
-        for (int a = 0; a < d.length; a++) {
-            if (a == 0) {
-                s = s + " " + d[a];
-            } else {
-                s = s + ", " + d[a];
-            }
-        }
-        return s + " ]";
-    }
-
-    public static String toString(BigDecimal[][] d) {
-        String s = "";
-        for (BigDecimal[] bigDecimals : d) {
-            s = s + toString(bigDecimals) + "\n";
-        }
-        return s.trim();
-    }
-
-    public static String toString(BigDecimal[] d) {
-        String s = "[";
-        for (int a = 0; a < d.length; a++) {
-            if (a == 0) {
-                s = s + " " + d[a];
-            } else {
-                s = s + ", " + d[a];
-            }
-        }
-        return s + " ]";
-    }
-
     /**
      * Given a matrix of variable coefficients for a linear system of equations,
      * this will solve for each variable.
@@ -277,13 +237,11 @@ public class Equations {
             row = row % coefficients.length;
             if (errorCounter > coefficients.length) {
                 if (VERBOSE_EXCEPTIONS) {
-                    throw new IllegalArgumentException("The coefficient matrix cannot be solved.  Either it has infinitely many solutions, or zero solutions:\n" + toString(coefficients));
+                    throw new IllegalArgumentException("The coefficient matrix cannot be solved.  Either it has infinitely many solutions, or zero solutions:\n" + MatrixToStringConverter.toString(coefficients));
                 } else {
                     throw new IllegalArgumentException("The coefficient matrix cannot be solved.  Either it has infinitely many solutions, or zero solutions.");
                 }
             }
-            //System.out.println("\trow = "+row);
-            //println(coefficients);
         }
 
         //System.out.println("\t\tmoving on...");
@@ -311,8 +269,6 @@ public class Equations {
                 coefficients[row][coefficients[row].length - 1] -= t;
                 coefficients[row][i] = 0;
             }
-            //System.out.println("\trow = "+row);
-            //println(coefficients);
         }
         if (sort) {
             Arrays.sort(coefficients, Equations::compareCoeffs);
@@ -415,7 +371,7 @@ public class Equations {
             row = row % coefficients.length;
             if (errorCounter > coefficients.length) {
                 if (VERBOSE_EXCEPTIONS) {
-                    throw new IllegalArgumentException("The coefficient matrix cannot be solved.  Either it has infinitely many solutions, or zero solutions:\n" + toString(coefficients));
+                    throw new IllegalArgumentException("The coefficient matrix cannot be solved.  Either it has infinitely many solutions, or zero solutions:\n" + MatrixToStringConverter.toString(coefficients));
                 } else {
                     throw new IllegalArgumentException("The coefficient matrix cannot be solved.  Either it has infinitely many solutions, or zero solutions.");
                 }
@@ -450,8 +406,6 @@ public class Equations {
                         .subtract(t);
                 coefficients[row][i] = BigDecimal.ZERO;
             }
-            //System.out.println("\trow = "+row);
-            //println(coefficients);
         }
         if (sort) {
             Arrays.sort(coefficients, Equations::compareBigCoeffs);
